+++ design/classdia_test.go
@@ -30,7 +30,7 @@ func exampleClassDiagram() *ClassDiagram {
 	d.HideRealizations()
 
 	var (
-		fnt      = d.Struct(shape.Font{})
+		fnt      = d.Struct(draw.Font{})
 		style    = d.Struct(draw.Style{})
 		seqdia   = d.Struct(SequenceDiagram{})
 		classdia = d.Struct(ClassDiagram{})