+++ shape/internet.go
@@ -3,6 +3,7 @@ package shape
 import (
 	"io"
 
+	"github.com/gregoryv/draw"
 	"github.com/gregoryv/nexus"
 )
 
@@ -10,8 +11,8 @@ func NewInternet() *Internet {
 	return &Internet{
 		Circle: Circle{Radius: 40, class: "internet"},
 		Title:  "Internet",
-		Font:   DefaultFont,
-		Pad:    DefaultTextPad,
+		Font:   draw.DefaultFont,
+		Pad:    draw.DefaultTextPad,
 		class:  "internet",
 	}
 }
@@ -20,8 +21,8 @@ type Internet struct {
 	Circle
 	Title string
 
-	Font  Font
-	Pad   Padding
+	Font  draw.Font
+	Pad   draw.Padding
 	class string
 }
 
@@ -39,12 +40,12 @@ func (r *Internet) title() *Label {
 		Text:  r.Title,
 		class: "internet-title",
 	}
-	label.Font.LineHeight = DefaultFont.Height
+	label.Font.LineHeight = draw.DefaultFont.Height
 	align := Aligner{}
 	align.VAlignCenter(&r.Circle, label)
 	align.HAlignCenter(&r.Circle, label)
 	return label
 }
 
-func (r *Internet) SetFont(f Font)         { r.Font = f }
-func (r *Internet) SetTextPad(pad Padding) { r.Pad = pad }
+func (r *Internet) SetFont(f draw.Font)         { r.Font = f }
+func (r *Internet) SetTextPad(pad draw.Padding) { r.Pad = pad }