+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [unreleased]
 
+- Remove aliased types and defaults in package shape
 - Hide field Diagram.Caption
 
 ## [0.30.0] 2023-08-02